// src/app/types/FormData.ts
export interface FormData {
  id: any;
  fname: string;
  email: string;
  phone: string;
  website: string;
  contactName: string;
  contactPhone: string;
  contactEmail: string;
  notes: string;
  type: string;
  category: string;
  comissionPercentage: number;
  activeForm: string;
  logo: string;
  criticalAccount: string;
  paymentOptions: string[];
}